committing sum mixed list

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python3
 """ This module contains a function that takes a list of floats and integers and returns their sum as a float """
-
 from typing import List, Union
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """ Returns the sum of the list of floats """
-    return sum(mxd_lst)    
+    return sum(mxd_lst)
